# Conflicts:
#	server/src/main/java/com/pfe/server/Models/User.java
#	server/src/main/java/com/pfe/server/Repositories/OrdersRepository.java
#	server/src/main/java/com/pfe/server/Repositories/ProductRepository.java
#	server/src/main/java/com/pfe/server/Repositories/UserRepository.java
#	server/src/main/java/com/pfe/server/Services/CategoriesService.java

+++ server/src/main/java/com/pfe/server/Controllers/ProfileController.java
@@ -56,7 +56,10 @@ public class ProfileController {
         }
         String oldImage = profile.getImage();
         profile.setImage(imageService.uploadImage(image));
-        imageService.deleteImage(oldImage);
+        //do not delete the default image
+        if (!oldImage.equals(imageService.getDefaultImage())) {
+            imageService.deleteImage(oldImage);
+        }
         return ResponseEntity.ok(profileService.updateProfile(profile));
     }
